+++ player/src/mcts/value.rs
@@ -1,7 +1,7 @@
 use game::*;
 
 #[derive(Debug, Clone, Copy, Default)]
-pub struct Value([f32; NUM_PLAYERS]);
+pub struct Value([f64; NUM_PLAYERS]);
 
 impl Value {
     pub fn from_game_scores(game_scores: [i16; NUM_PLAYERS]) -> Self {
@@ -12,20 +12,20 @@ impl Value {
         if score_range == 0 {
             // If all scores are the same, return 1 / NUM_PLAYERS for each player
             // e.g. if there are 2 players, return [0.5, 0.5] for each player
-            return Self([1.0 / NUM_PLAYERS as f32; NUM_PLAYERS]);
+            return Self([1.0 / NUM_PLAYERS as f64; NUM_PLAYERS]);
         }
 
         let mut value = [0.0; NUM_PLAYERS];
-        let score_range = score_range as f32;
+        let score_range = score_range as f64;
         for (i, &score) in game_scores.iter().enumerate() {
-            let normalized_score = (score - min_score) as f32 / score_range;
+            let normalized_score = (score - min_score) as f64 / score_range;
             // Add very small bonus for higher scores.
             // Otherwise the program will do random moves as soon as its victory is inevitable which is perceived as arrogant by other players.
-            value[i] = normalized_score + score as f32 * 0.001;
+            value[i] = normalized_score + score as f64 * 0.001;
         }
 
         // Divide by the sum of all values to normalize them
-        let sum: f32 = value.iter().sum();
+        let sum: f64 = value.iter().sum();
         for value in value.iter_mut() {
             *value /= sum;
         }
@@ -55,39 +55,39 @@ impl std::ops::AddAssign for Value {
     }
 }
 
-impl std::ops::DivAssign<f32> for Value {
-    fn div_assign(&mut self, rhs: f32) {
+impl std::ops::DivAssign<f64> for Value {
+    fn div_assign(&mut self, rhs: f64) {
         for value in self.0.iter_mut() {
             *value /= rhs;
         }
     }
 }
 
-impl std::ops::Div<f32> for Value {
+impl std::ops::Div<f64> for Value {
     type Output = Self;
 
-    fn div(mut self, rhs: f32) -> Self::Output {
+    fn div(mut self, rhs: f64) -> Self::Output {
         self /= rhs;
         self
     }
 }
 
 impl std::ops::Index<usize> for Value {
-    type Output = f32;
+    type Output = f64;
 
     fn index(&self, index: usize) -> &Self::Output {
         &self.0[index]
     }
 }
 
-impl std::convert::From<Value> for [f32; NUM_PLAYERS] {
+impl std::convert::From<Value> for [f64; NUM_PLAYERS] {
     fn from(value: Value) -> Self {
         value.0
     }
 }
 
-impl std::convert::From<[f32; NUM_PLAYERS]> for Value {
-    fn from(value: [f32; NUM_PLAYERS]) -> Self {
+impl std::convert::From<[f64; NUM_PLAYERS]> for Value {
+    fn from(value: [f64; NUM_PLAYERS]) -> Self {
         Self(value)
     }
 }
